@@ -6,7 +6,7 @@ authors = ["Thomas Oberliessen <thomas.oberliessen@googlemail.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">3.10, <3.12"
+python = ">3.11, <3.12"
 pandas = "2.0.1"
 jupyter = "^1.0.0"
 matplotlib = "^3.6.0"
