@@ -134,10 +134,7 @@ class SystemParticipantsContainer(ContainerMixin):
 
     def uuids(self):
         return pd.concat(
-            [
-                participants.uuid.to_series()
-                for participants in self.to_list(include_empty=True)
-            ]
+            [participants.uuid for participants in self.to_list(include_empty=True)]
         )
 
     def subset(self, uuids):
